# packages/api-schema/openapi/core.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PERUM Control Plane",
    "description": "Manages organizations, provisioning, billing and observability for PERUM tenant stacks.",
    "version": "0.5.3"
  },
  "paths": {
...
    "/internal/support/escalations": {
      "post": {
        "tags": [
          "internal-support"
        ],
        "summary": "Intake Escalation",
        "operationId": "intake_escalation_internal_support_escalations_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationIntake"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Intake Escalation Internal Support Escalations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/support/escalations/outbound": {
      "get": {
        "tags": [
          "internal-support"
        ],
        "summary": "Outbound Escalation",
        "operationId": "outbound_escalation_internal_support_escalations_outbound_get",
        "parameters": [
          {
            "name": "school_public_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "School Public Id"
            }
          },
          {
            "name": "correlation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Correlation Id"
            }
          },
          {
            "name": "since_cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Since Cursor"
            }
          },
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Outbound Escalation Internal Support Escalations Outbound Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/support/escalations/outbound/ack": {
      "post": {
        "tags": [
          "internal-support"
        ],
        "summary": "Ack Outbound",
        "operationId": "ack_outbound_internal_support_escalations_outbound_ack_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundAck"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ack Outbound Internal Support Escalations Outbound Ack Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/support/tickets": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "My Tickets",
        "operationId": "my_tickets_api_support_tickets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response My Tickets Api Support Tickets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "support"
        ],
        "summary": "Open Ticket",
        "operationId": "open_ticket_api_support_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Open Ticket Api Support Tickets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/support/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "Org Ticket",
        "operationId": "org_ticket_api_support_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Org Ticket Api Support Tickets  Ticket Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/support/tickets/{ticket_id}/messages": {
      "post": {
        "tags": [
          "support"
        ],
        "summary": "Org Reply",
        "operationId": "org_reply_api_support_tickets__ticket_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Org Reply Api Support Tickets  Ticket Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/support/admin/badge": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "Support Badge",
        "operationId": "support_badge_api_support_admin_badge_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Support Badge Api Support Admin Badge Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/support/admin/tickets": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "All Tickets",
        "operationId": "all_tickets_api_support_admin_tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response All Tickets Api Support Admin Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/support/escalations/pending": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "Pending Escalations",
        "operationId": "pending_escalations_api_support_escalations_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Pending Escalations Api Support Escalations Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EscalationIntake": {
        "properties": {
          "school_public_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Public Id"
          },
          "correlation_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Correlation Id"
          },
          "tenant_ticket_public_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Ticket Public Id"
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Subject"
          },
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "client_message_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          },
          "redacted_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Redacted Snapshot"
          }
        },
        "type": "object",
        "required": [
          "school_public_id",
          "correlation_id",
          "subject",
          "message"
        ],
        "title": "EscalationIntake"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}